refactored upload component
the input is not used anymore and I added a button and a label text for
the component. Furthermore the callback is not optional anymore because
the component itself has no knowledge what it should do with the files
selected, so the callback is indeed required. Also the drag and click
events do not trigger a click event on the input element itself, this
caused recursive error in the console on the client so the open picker
is not triggered directly.